fix: shared typing on decrypted attachments
